TEJI Desk: multi-instrument, crypto (BTC/ETH), timeframes & in-app controls
Turns the single-instrument runner into a multi-market desk with runtime
controls, per the requested features:

- Multiple instruments at once: a watchlist (Nifty, Bank Nifty, a stock, BTC,
  ETH), each with its own Trader (independent candles, strategy state, position
  and P&L). Portfolio-level P&L and a shared daily-loss limit.
- Bitcoin & Ethereum in rupees via the CoinDCX public ticker — real live prices
  with NO account in paper mode (24x7). Live crypto orders left as a documented
  TODO (needs a signed exchange adapter).
- Selectable timeframe (1m/5m/15m/1h), changeable live from the UI; changing it
  rebuilds each instrument's candle aggregation and re-warms the strategy.
- In-app controls via a POST /api/config settings endpoint: per-instrument
  trading on/off toggles, trade-all/pause-all, timeframe switch, and flat-all
  (kill switch) — no restart, no file edits.
- Fractional quantities so crypto (e.g. 0.01 BTC) trades alongside equity lots.

Refactor: Feed.on_tick now carries the symbol; run.py builds one feed per source
(mock/crypto/angelone) each serving its subset; State is namespaced per symbol;
Engine doubles as the dashboard's runtime controller. Dashboard rebuilt with an
instrument rail (selectable, per-row toggle), header controls, and a per-instrument
detail view. Verified end-to-end on the mock feed: 5 instruments trading, enable/
disable and timeframe switch applied live, crypto priced and trading in rupees.

+++ teji-desk/teji/risk/manager.py
@@ -13,9 +13,9 @@ from ..data.candles import now_ist
 
 
 class RiskManager:
-    def __init__(self, cfg: dict, lot_size: int, kill_file: str = "KILL"):
-        self.qty_lots = int(cfg.get("qty_lots", 1))
-        self.lot_size = int(lot_size)
+    def __init__(self, cfg: dict, lot_size: float, kill_file: str = "KILL"):
+        self.qty_lots = float(cfg.get("qty_lots", 1))
+        self.lot_size = float(lot_size)   # float: crypto trades fractional units
         self.max_daily_loss = float(cfg.get("max_daily_loss", 3000))
         self.max_trades = int(cfg.get("max_trades", 8))
         self.square_off = str(cfg.get("square_off", "15:20"))
@@ -23,7 +23,8 @@ class RiskManager:
         self.kill_file = Path(kill_file)
 
     @property
-    def order_qty(self) -> int:
+    def order_qty(self) -> float:
+        # float so crypto (fractional units) works alongside equity lots
         return self.qty_lots * self.lot_size
 
     def kill_switch_engaged(self) -> bool:
